Change coords to f64 and add unit test
The API allows at most 6 digits after the decimal point. In testing I realized that the f32 data type is not always correctly converted to string literals in the format!() macro. This issue is resolved by using f64 data types for the coordinates.
Also I added a unit test for the coords_to_str() function.